feat(bdd): unify native and web testing with modular wdio
