+++ .gitignore
@@ -33,3 +33,4 @@ web-build/
 
 # Yarn 4+ internal files
 .yarn/install-state.gz
+.swp
