+++ apps/bdd/.env.example
@@ -1,0 +1,2 @@
+PROJECT_ROOT_ANDROID=
+PROJECT_ROOT_IOS=
