+++ apps/bdd/.gitignore
@@ -1,2 +1,4 @@
 node_modules
 test-artifacts/
+logs/
+.env
