+++ apps/bdd/features/iteration2/landing.feature
@@ -1,7 +1,6 @@
-@mobile_native @skip
-Feature: The first page of the native app
-
-  Scenario : As a user I can open the app
-    Given I am on the phone home screen
-    When I open the app
-    Then I should see the app's landing page
+@native 
+Feature: App Launch
+  Scenario: User opens the app for the first time
+    Given the app is installed
+    When I launch the app
+    Then I should see the first screen
