+++ apps/bdd/features/iteration5/buy_a_token_from_greenstand.feature
@@ -1,4 +1,4 @@
-@skip @web @mobile_native
+@skip @web @native
 Feature: buy a token from Greenstand
 
   Scenario: As a user, I can buy a token from Greenstand
