+++ apps/bdd/package.json
@@ -5,20 +5,22 @@
   "description": "WebdriverIO End-to-End Tests for Treetracker Wallet App",
   "main": "index.js",
   "scripts": {
-    "wdio": "wdio run ./wdio.conf.ts",
-    "wdio:watch": "wdio run ./wdio.conf.ts --watch",
+    "wdio": "wdio run ./wdio.web.conf.ts",
+    "wdio:watch": "wdio run ./wdio.web.conf.ts --watch",
     "type-check": "tsc --noEmit",
     "build": "tsc",
-    "test": "wdio run ./wdio.conf.ts",
-    "test:local": "wdio run ./wdio.conf.ts --baseUrl http://localhost:3000",
-    "test:headless": "wdio run ./wdio.conf.ts --headless",
-    "test:login": "wdio run ./wdio.conf.ts --spec ./features/login.feature",
-    "test:register": "wdio run ./wdio.conf.ts --spec ./features/register.feature",
-    "test:debug": "wdio run ./wdio.conf.ts --logLevel debug",
+    "test:web": "cross-env wdio run ./wdio.web.conf.ts",
+    "test:web:local": "cross-env wdio run ./wdio.web.conf.ts --baseUrl http://localhost:3000",
+    "test:web:headless": "cross-env wdio run ./wdio.web.conf.ts --headless",
+    "test:web:login": "cross-env wdio run ./wdio.web.conf.ts --spec ./features/login.feature",
+    "test:web:register": "cross-env wdio run ./wdio.web.conf.ts --spec ./features/register.feature",
+    "test:web:debug": "cross-env wdio run ./wdio.debug.conf.ts --logLevel debug",
     "test:update-driver": "yarn add -D chromedriver@latest",
     "report:cucumber": "node scripts/generate-cucumber-report.js",
     "test:report": "node scripts/bdd-test-report.js",
-    "test:local:report": "node scripts/bdd-test-report.js test:local"
+    "test:local:report": "node scripts/bdd-test-report.js test:local",
+    "test:native:android": "cross-env PLATFORM=android wdio run ./wdio.mobile.conf.ts",
+    "test:native:ios": "cross-env PLATFORM=ios wdio run ./wdio.mobile.conf.ts"
   },
   "devDependencies": {
     "@badisi/wdio-harness": "^4.2.0",
@@ -31,7 +33,10 @@
     "@wdio/jasmine-framework": "^9.18.0",
     "@wdio/local-runner": "^9.18.4",
     "@wdio/spec-reporter": "^9.18.0",
-    "chromedriver": "^141.0.0",
+    "chromedriver": "^143.0.1",
+    "cross-env": "^10.1.0",
+    "dotenv": "^17.2.3",
+    "msw": "^2.11.6",
     "multiple-cucumber-html-reporter": "^3.9.3",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.2",
