+++ apps/bdd/tsconfig.json
@@ -14,6 +14,13 @@
       "@wdio/globals": ["node_modules/@wdio/globals/types"]
     }
   },
-  "include": ["**/*.ts", ".wdio.conf.ts", "features/**/*.ts"],
+  "include": [
+    "**/*.ts",
+    ".wdio.conf.ts",
+    "features/**/*.ts",
+    "mocks/*.ts",
+    "support/*.ts",
+    "mocks/hooks"
+  ],
   "exclude": ["node_modules"]
 }
