+++ apps/native/.gitignore
@@ -24,3 +24,5 @@ logs
 
 android 
 ios 
+
+tsconfig.tsbuildinfo
