+++ apps/native/package.json
@@ -27,7 +27,9 @@
     "pre-commit": "lint-staged",
     "prebuild": "expo prebuild",
     "prebuild-clean": "expo prebuild --clean",
-    "build-android": "cd android && ./gradlew assembleDebug"
+    "build-android": "cd android && ./gradlew assembleDebug",
+    "build-ios": "cd ios && xcodebuild -workspace native.xcworkspace -scheme native -sdk iphonesimulator -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 17 Pro,OS=latest'",
+    "install-android-debug": "cd android && ./gradlew installDebug"
   },
   "jest": {
     "preset": "jest-expo"
@@ -38,13 +40,13 @@
     "@radix-ui/react-compose-refs": "^1.1.2",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-native/virtualized-lists": "^0.77.1",
-    "@react-navigation/bottom-tabs": "^7.8.5",
+    "@react-navigation/bottom-tabs": "^7.4.0",
     "@react-navigation/native": "^7.0.14",
     "@shopify/flash-list": "2.0.2",
     "add": "^2.0.6",
     "chai": "^6.2.0",
     "core": "1.0.0",
-    "expo": "^54.0.18",
+    "expo": "~54.0.25",
     "expo-build-properties": "^1.0.9",
     "expo-constants": "~18.0.9",
     "expo-dev-client": "^6.0.16",
@@ -60,7 +62,7 @@
     "react": "19.1.0",
     "react-compiler-runtime": "^19.1.0-rc.2",
     "react-dom": "19.1.0",
-    "react-native": "0.81.4",
+    "react-native": "0.81.5",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-paper": "^5.12.3",
     "react-native-reanimated": "~4.1.1",
