--- apps/native/tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./declarations.d.ts","./wdio.conf.ts","./constants/colors.ts","./e2e/features/step-definitions/app.steps.ts","./e2e/mocks/handlers.ts","./e2e/mocks/server.ts","./hooks/usecolorscheme.ts","./hooks/usethemecolor.ts","./utils/dimensions.ts","./app/_layout.tsx","./app/accountconfirmation.tsx","./app/index.tsx","./app/(auth)/checkemail.tsx","./app/(auth)/forgotpassword.tsx","./app/(auth)/login.tsx","./app/(tabs)/_layout.tsx","./app/(tabs)/home/_layout.tsx","./app/(tabs)/home/index.tsx","./app/(tabs)/notifications/_layout.tsx","./app/(tabs)/notifications/index.tsx","./app/(tabs)/settings/_layout.tsx","./app/(tabs)/settings/index.tsx","./app/(tabs)/wallet/_layout.tsx","./app/(tabs)/wallet/index.tsx","./app/onboarding/index.tsx","./components/actionselectionmodal.tsx","./components/button.tsx","./components/collapsible.tsx","./components/customtabbar.tsx","./components/externallink.tsx","./components/headersearch.tsx","./components/parallaxscrollview.tsx","./components/themesafearea.tsx","./components/themedsafearea.tsx","./components/themedstatusbar.tsx","./components/themedtext.tsx","./components/themedview.tsx","./components/__tests__/themedtext-test.tsx","./components/navigation/tabbaricon.tsx","./components/svg/cloud.tsx","./components/svg/leaf.tsx","./components/svg/wallet.tsx","./components/ui/balancecard.tsx","./components/ui/common/custombutton.tsx","./components/ui/common/custommodal.tsx","./components/ui/common/customsubmitbutton.tsx","./components/ui/common/customtextinput.tsx","./components/ui/common/customtitle.tsx","./components/ui/common/socialloginbutton.tsx","./components/wallet/walletlist.tsx","./components/wallet/yourwallets.tsx","./context/modalcontext.tsx"],"version":"5.9.3"}
