+++ package.json
@@ -10,26 +10,26 @@
   "type": "module",
   "scripts": {
     "type-check": "tsc --build tsconfig.json --noEmit",
-    "web:dev": "yarn workspace web dev",
-    "web:prod": "yarn workspace web start",
-    "web:build": "yarn workspace build",
-    "native:start": "yarn workspace native start",
-    "native:android": "yarn workspace native android",
-    "native:ios": "yarn workspace native ios",
-    "user:build": "yarn workspace user build",
-    "user:dev": "yarn workspace user start:dev",
-    "user:debug": "yarn workspace user start:debug",
-    "user:prod": "yarn workspace user start:prod",
-    "user:test:unit": "yarn workspace user test:unit",
-    "user:test:unit:watch": "yarn workspace user test:watch",
-    "user:test:int": "yarn workspace user int-test",
-    "user:test:cov": "yarn workspace user test:cov",
-    "user:test:debug": "yarn workspace user test:debug",
-    "user:test:e2e": "yarn workspace user test:e2e",
+    "web:dev": "yarn workspace web  run dev",
+    "web:prod": "yarn workspace web run start",
+    "web:build": "yarn workspace web run build",
+    "native:start": "yarn workspace native run start",
+    "native:android": "yarn workspace native run android",
+    "native:ios": "yarn workspace native run ios",
+    "user:build": "yarn workspace user run build",
+    "user:dev": "yarn workspace user run start:dev",
+    "user:debug": "yarn workspace user run start:debug",
+    "user:prod": "yarn workspace user run start:prod",
+    "user:test:unit": "yarn workspace user run test:unit",
+    "user:test:unit:watch": "yarn workspace user run test:watch",
+    "user:test:int": "yarn workspace user run int-test",
+    "user:test:cov": "yarn workspace user run test:cov",
+    "user:test:debug": "yarn workspace user run test:debug",
+    "user:test:e2e": "yarn workspace user run test:e2e",
     "pre-commit": "lint-staged",
     "cypress-component-test": "yarn workspace web run cy:component",
-    "cypress-e2e-test": "yarn workspace web cy:e2e",
-    "cypress-e2e-headless-test": "yarn workspace web cy:e2e-headless",
+    "cypress-e2e-test": "yarn workspace web run cy:e2e",
+    "cypress-e2e-headless-test": "yarn workspace web run cy:e2e-headless",
     "commitlint": "commitlint",
     "lint-staged": "lint-staged",
     "lint:check": "eslint . --no-warn-ignored",
@@ -38,18 +38,22 @@
     "format:check": "prettier --config packages/config/prettier.config.json --check .",
     "postinstall": "husky",
     "prepare": "husky",
-    "bdd:e2e": "yarn workspace @treetracker/bdd test",
-    "bdd:e2e:local": "yarn workspace @treetracker/bdd test:local",
-    "bdd:e2e:login": "yarn workspace @treetracker/bdd test:login",
-    "bdd:e2e:register": "yarn workspace @treetracker/bdd test:register",
+    "bdd:native": "yarn workspace @treetracker/bdd run test:native",
+    "bdd:e2e": "yarn workspace @treetracker/bdd run test:web",
+    "bdd:e2e:local": "yarn workspace @treetracker/bdd run test:local",
+    "bdd:e2e:login": "yarn workspace @treetracker/bdd run  test:login",
+    "bdd:e2e:register": "yarn workspace @treetracker/bdd run  test:register",
     "bdd:e2e:debug": "yarn workspace @treetracker/bdd test:debug",
-    "bdd:e2e:update-driver": "yarn workspace @treetracker/bdd test:update-driver",
-    "bdd:report": "yarn workspace @treetracker/bdd report:cucumber",
-    "bdd:test:report": "yarn workspace @treetracker/bdd test:report",
-    "bdd:test:local:report": "yarn workspace @treetracker/bdd test:local:report",
-    "native:bdd:test": "yarn workspace native wdio",
+    "bdd:e2e:update-driver": "yarn workspace @treetracker/bdd run  test:update-driver",
+    "bdd:report": "yarn workspace @treetracker/bdd  run report:cucumber",
+    "bdd:test:report": "yarn workspace @treetracker/bdd run  test:report",
+    "bdd:test:local:report": "yarn workspace @treetracker/bdd run  test:local:report",
+    "bdd:test:android": "yarn workspace @treetracker/bdd run test:native:android",
+    "bdd:test:ios": "yarn workspace @treetracker/bdd run  test:native:ios",
+    "native:bdd:test": "yarn workspace native run  wdio",
     "build-apk-android": "yarn workspace native run build-android",
-    "prebuild": "yarn workspace native prebuild"
+    "prebuild": "yarn workspace native run prebuild",
+    "install-android-debug": "yarn workspace native run install-android-debug"
   },
   "workspaces": {
     "packages": [
